aula2 points controller
